[V2] Code Conformance Cleanup (BREAKING CHANGES) (#76)
* Linting Cleanup and snake_case Naming Cleanup
Rename message functions to be more descriptive.

Rename _upmsg, _routemsg, _controlmsg to more descriptive names.
Fix Error Types for XML Parsing


Add missing IndexError

* Rename nid, pid to address for consistency. Rename on/off to turn_on/off

* Linting mopup, Add isy.connected property


Fix Line Endings.

* Rename folders to lowercase for conformance.

* Fix new capitalizations

Remaining PyLint Report:

```
************* Module PyISY.climate
PyISY/climate.py:246:8: W0702: No exception type(s) specified (bare-except)
PyISY/climate.py:259:16: W0702: No exception type(s) specified (bare-except)
PyISY/climate.py:299:4: C0116: Missing function or method docstring (missing-function-docstring)
************* Module PyISY.connection
PyISY/connection.py:17:0: E0401: Unable to import 'requests.packages.urllib3.poolmanager' (import-error)
PyISY/connection.py:61:16: E1101: Module 'requests.packages' has no 'urllib3' member (no-member)
PyISY/connection.py:91:38: W1505: Using deprecated method encodestring() (deprecated-method)
************* Module PyISY
PyISY/__init__.py:1:0: C0103: Module name "PyISY" doesn't conform to snake_case naming style (invalid-name)
PyISY/__init__.py:30:7: W0703: Catching too general exception Exception (broad-except)
************* Module PyISY.nodes.node
PyISY/nodes/node.py:223:-1: W0105: String statement has no effect (pointless-string-statement)
************* Module PyISY.nodes.nodebase
PyISY/nodes/nodebase.py:138:15: W0212: Access to a protected member _val of a client class (protected-access)
************* Module PyISY.programs.__init__
PyISY/programs/__init__.py:244:2: W0511: TODO: Unpack into proper properties. (fixme)
```

* Fix missing Error Type

* Add pylintrc file

* Corrections to capitalization and update CHANGELOG

* Minor pylint corrections.
