+++ pyisy/nodes/__init__.py
@@ -55,12 +55,12 @@ class Nodes:
 
     This class can be used as a dictionary to
     navigate through the controller's structure to objects of type
-    :class:`~PyISY.Nodes.Node` and :class:`~PyISY.Nodes.Group` that represent
+    :class:`pyisy.nodes.Node` and :class:`pyisy.nodes.Group` that represent
     objects on the controller.
 
     |  isy: ISY class
     |  root: [optional] String representing the current navigation level's ID
-    |  nids: [optional] list of node ids
+    |  addresses: [optional] list of node ids
     |  nnames: [optional] list of node names
     |  nparents: [optional] list of node parents
     |  nobjs: [optional] list of node objects
@@ -69,11 +69,11 @@ class Nodes:
 
     :ivar all_lower_nodes: Return all nodes beneath current level
     :ivar children: A list of the object's children.
-    :ivar hasChildren: Indicates if object has children
+    :ivar has_children: Indicates if object has children
     :ivar name: The name of the current folder in navigation.
     """
 
-    nids = []
+    addresses = []
     nnames = []
     nparents = []
     nobjs = []
@@ -83,7 +83,7 @@ class Nodes:
         self,
         isy,
         root=None,
-        nids=None,
+        addresses=None,
         nnames=None,
         nparents=None,
         nobjs=None,
@@ -95,14 +95,14 @@ class Nodes:
         self.root = root
 
         if (
-            nids is not None
+            addresses is not None
             and nnames is not None
             and nparents is not None
             and nobjs is not None
             and ntypes is not None
         ):
 
-            self.nids = nids
+            self.addresses = addresses
             self.nnames = nnames
             self.nparents = nparents
             self.nobjs = nobjs
@@ -115,7 +115,7 @@ class Nodes:
         """Return string representation of the nodes/folders/groups."""
         if self.root is None:
             return "Folder <root>"
-        ind = self.nids.index(self.root)
+        ind = self.addresses.index(self.root)
         if self.ntypes[ind] == TAG_FOLDER:
             return "Folder ({})".format(self.root)
         if self.ntypes[ind] == TAG_GROUP:
@@ -172,12 +172,12 @@ class Nodes:
         out = ""
         for node in nodes:
             node_obj = self[node[2]]
-            if node_obj.hasChildren:
+            if node_obj.has_children:
                 out += "  + "
             else:
                 out += "  "
             out += "{}: Node({})\n".format(node[1], node[2])
-            if node_obj.hasChildren:
+            if node_obj.has_children:
                 for line in repr(node_obj).split("\n")[1:]:
                     out += "  |   {}\n".format(line)
                 out += "  -\n"
@@ -195,18 +195,18 @@ class Nodes:
 
     def update_received(self, xmldoc):
         """Update nodes from event stream message."""
-        nid = value_from_xml(xmldoc, TAG_NODE)
+        address = value_from_xml(xmldoc, TAG_NODE)
         nval = int(value_from_xml(xmldoc, ATTR_ACTION))
         prec = attr_from_xml(xmldoc, ATTR_ACTION, ATTR_PRECISION, "0")
         uom = attr_from_xml(xmldoc, ATTR_ACTION, ATTR_UNIT_OF_MEASURE, "")
-        node = self.get_by_id(nid)
+        node = self.get_by_id(address)
+        if not node:
+            return
         # Check if UOM/PREC have changed or were not set:
-        if prec and prec != node.prec:
-            node.prec = prec
-        if uom and uom != node.uom:
-            node.uom = uom
-        self.get_by_id(nid).status.update(nval, force=True, silent=True)
-        self.isy.log.debug("ISY Updated Node: " + nid)
+        node.update_precision(prec)
+        node.update_uom(uom)
+        node.status.update(nval, force=True, silent=True)
+        self.isy.log.debug("ISY Updated Node: " + address)
 
     def control_message_received(self, xmldoc):
         """
@@ -214,9 +214,9 @@ class Nodes:
 
         Used for sending out to subscribers.
         """
-        nid = value_from_xml(xmldoc, TAG_NODE)
+        address = value_from_xml(xmldoc, TAG_NODE)
         cntrl = value_from_xml(xmldoc, ATTR_CONTROL)
-        if not (nid and cntrl):
+        if not (address and cntrl):
             # If there is no node associated with the control message ignore it
             return
 
@@ -226,20 +226,20 @@ class Nodes:
         uom = attr_from_xml(xmldoc, ATTR_ACTION, ATTR_UNIT_OF_MEASURE, "")
         formatted = attr_from_xml(xmldoc, ATTR_FORMATTED, nval)
 
-        node = self.get_by_id(nid)
+        node = self.get_by_id(address)
         if not node:
             self.isy.log.debug(
                 "Received a node update for node %s but could not find a record of this "
                 "node. Please try restarting the module if the problem persists, this "
                 "may be due to a new node being added to the ISY since last restart.",
-                nid,
+                address,
             )
             return
 
         if cntrl == PROP_RAMP_RATE:
             nval = INSTEON_RAMP_RATES.get(nval, nval)
         if cntrl not in EVENT_PROPS_IGNORED:
-            node._aux_properties[cntrl] = {
+            node.aux_properties[cntrl] = {
                 ATTR_ID: cntrl,
                 ATTR_VALUE: nval,
                 ATTR_PRECISION: prec,
@@ -247,7 +247,7 @@ class Nodes:
                 ATTR_FORMATTED: formatted,
             }
         node.control_events.notify(EventResult(cntrl, nval, prec, uom, formatted))
-        self.isy.log.debug("ISY Node Control Event: %s %s %s", nid, cntrl, nval)
+        self.isy.log.debug("ISY Node Control Event: %s %s %s", address, cntrl, nval)
 
     def parse(self, xml):
         """
@@ -257,7 +257,7 @@ class Nodes:
         """
         try:
             xmldoc = minidom.parseString(xml)
-        except:
+        except (AttributeError, KeyError, ValueError, TypeError, IndexError):
             self.isy.log.error("%s: Nodes", XML_PARSE_ERROR)
             return False
 
@@ -268,7 +268,7 @@ class Nodes:
 
             for feature in features:
                 # Get Node Information
-                nid = value_from_xml(feature, TAG_ADDRESS)
+                address = value_from_xml(feature, TAG_ADDRESS)
                 nname = value_from_xml(feature, TAG_NAME)
                 nparent = value_from_xml(feature, TAG_PARENT)
                 parent_nid = value_from_xml(feature, TAG_PRIMARY_NODE)
@@ -301,20 +301,20 @@ class Nodes:
                             protocol = f"{PROTO_NODE_SERVER}_{node_server}"
 
                 # Process the different node types
-                if ntype == TAG_FOLDER and nid not in self.nids:
-                    self.insert(nid, nname, nparent, None, ntype)
+                if ntype == TAG_FOLDER and address not in self.addresses:
+                    self.insert(address, nname, nparent, None, ntype)
                 elif ntype == TAG_NODE:
-                    if nid in self.nids:
-                        self.get_by_id(nid).update(xmldoc=feature)
+                    if address in self.addresses:
+                        self.get_by_id(address).update(xmldoc=feature)
                         continue
                     state, aux_props = parse_xml_properties(feature)
                     self.insert(
-                        nid,
+                        address,
                         nname,
                         nparent,
                         Node(
                             self,
-                            nid=nid,
+                            address=address,
                             name=nname,
                             state=state,
                             aux_properties=aux_props,
@@ -329,7 +329,7 @@ class Nodes:
                         ),
                         ntype,
                     )
-                elif ntype == TAG_GROUP and nid not in self.nids:
+                elif ntype == TAG_GROUP and address not in self.addresses:
                     flag = attr_from_element(feature, ATTR_FLAG)
                     # Ignore groups that contain 0x08 in the flag since
                     # that is a ISY scene that contains every device/
@@ -338,7 +338,7 @@ class Nodes:
                     # the ISY MAC addrees in newer versions of
                     # ISY firmwares > 5.0.6+ ..
                     if int(flag) & 0x08:
-                        self.isy.log.info("Skipping group flag=%s %s", flag, nid)
+                        self.isy.log.info("Skipping group flag=%s %s", flag, address)
                         continue
                     mems = feature.getElementsByTagName(TAG_LINK)
                     # Build list of members
@@ -349,10 +349,10 @@ class Nodes:
                         if int(attr_from_element(mem, TAG_TYPE, 0)) == 16:
                             controllers.append(mem.firstChild.nodeValue)
                     self.insert(
-                        nid,
+                        address,
                         nname,
                         nparent,
-                        Group(self, nid, nname, members, controllers),
+                        Group(self, address, nname, members, controllers),
                         ntype,
                     )
             self.isy.log.debug("ISY Loaded {}".format(ntype))
@@ -370,17 +370,17 @@ class Nodes:
         else:
             self.isy.log.warning("ISY Failed to update nodes.")
 
-    def insert(self, nid, nname, nparent, nobj, ntype):
+    def insert(self, address, nname, nparent, nobj, ntype):
         """
         Insert a new node into the lists.
 
-        |  nid: node id
+        |  address: node id
         |  nname: node name
         |  nparent: node parent
         |  nobj: node object
         |  ntype: node type
         """
-        self.nids.append(nid)
+        self.addresses.append(address)
         self.nnames.append(nname)
         self.nparents.append(nparent)
         self.ntypes.append(ntype)
@@ -389,7 +389,7 @@ class Nodes:
     def __getitem__(self, val):
         """Navigate through the node tree. Can take names or IDs."""
         try:
-            self.nids.index(val)
+            self.addresses.index(val)
             fun = self.get_by_id
         except ValueError:
             try:
@@ -406,7 +406,7 @@ class Nodes:
             output = None
             try:
                 output = fun(val)
-            except:
+            except ValueError:
                 pass
 
             if output:
@@ -423,19 +423,19 @@ class Nodes:
 
         |  val: String representing name to look for.
         """
-        for i in range(len(self.nids)):
+        for i in range(len(self.addresses)):
             if self.nparents[i] == self.root and self.nnames[i] == val:
                 return self.get_by_index(i)
         return None
 
-    def get_by_id(self, nid):
+    def get_by_id(self, address):
         """
         Get object with the given ID.
 
-        |  nid: Integer representing node/group/folder id.
+        |  address: Integer representing node/group/folder id.
         """
         try:
-            i = self.nids.index(nid)
+            i = self.addresses.index(address)
         except ValueError:
             return None
         else:
@@ -451,8 +451,8 @@ class Nodes:
             return self.nobjs[i]
         return Nodes(
             self.isy,
-            self.nids[i],
-            self.nids,
+            self.addresses[i],
+            self.addresses,
             self.nnames,
             self.nparents,
             self.nobjs,
@@ -463,18 +463,18 @@ class Nodes:
     def children(self):
         """Return the children of the class."""
         out = []
-        for i in range(len(self.nids)):
+        for i in range(len(self.addresses)):
             if self.nparents[i] == self.root:
-                out.append((self.ntypes[i], self.nnames[i], self.nids[i]))
+                out.append((self.ntypes[i], self.nnames[i], self.addresses[i]))
         return out
 
     @property
-    def hasChildren(self):
+    def has_children(self):
         """Return if the root has children."""
         try:
             self.nparents.index(self.root)
             return True
-        except:
+        except ValueError:
             return False
 
     @property
@@ -482,7 +482,7 @@ class Nodes:
         """Return the name of the root."""
         if self.root is None:
             return ""
-        ind = self.nids.index(self.root)
+        ind = self.addresses.index(self.root)
         return self.nnames[ind]
 
     @property
